Reproduce their output-head init and make the network boundary fp32-safe
Their NormalPolicy/NormalValue run init_fc_weights on both heads (weights x0.1,
bias 0). That is why their iter-0 eval reward is ~499: an untrained net emits
near-zero mean actions, the ant stands still, and the episode is 500 x the +1
survive bonus. With torch's default init the head emits ~0.1, which at gear 150
is real torque, and the untrained eval lands ~40-60 lower. Also cast obs at the
network boundary so the float64 CPU (parity) backend can drive PPO, and check
the reset state distribution against theirs in the gate.

+++ rower_soccer/competevo_port/render.py
@@ -54,8 +54,8 @@ def eval_video(env, ac, path, renderer=None, fps=40, world=0, max_steps=None):
     frames, rew_sum, winner = [], np.zeros(env.n_agents), None
     steps = max_steps or env.max_episode_steps
     for t in range(steps):
-        a = ac.mean_action(obs.reshape(-1, env.obs_dim)).clamp(-1, 1)
-        obs, rew, done, info = env.step(a.reshape(env.n, env.n_agents, -1))
+        a = ac.mean_action(obs.reshape(-1, env.obs_dim).float()).clamp(-1, 1)
+        obs, rew, done, info = env.step(a.reshape(env.n, env.n_agents, -1).to(env.dtype))
         rew_sum += rew[world].float().cpu().numpy()
         frames.append(renderer.frame(env, w=world))
         if bool(done[world]):
